gcc: silence alloc warning
This commit silences the following gcc warning:
gdi/gpixmap.cpp:987:48: warning: argument 1 value ‘18446744073709551615’ exceeds maximum object size 9223372036854775807 [-Walloc-size-larger-than=]
gdi/grc.cpp:444:22: warning: argument 1 value ‘18446744073709551615’ exceeds maximum object size 9223372036854775807 [-Walloc-size-larger-than=]

Everything that used on malloc/calloc/realloc/new should be an unsigned/size_t type.

+++ lib/gdi/gpixmap.cpp
@@ -984,7 +984,7 @@ void gPixmap::mergePalette(const gPixmap &target)
 	if ((!surface->clut.colors) || (!target.surface->clut.colors))
 		return;
 
-	gColor *lookup=new gColor[surface->clut.colors];
+	gColor *lookup=new gColor[static_cast<size_t>(surface->clut.colors)];
 
 	for (int i=0; i<surface->clut.colors; i++)
 		lookup[i].color=target.surface->clut.findColor(surface->clut.data[i]);
