+++ lib/gdi/grc.cpp
@@ -441,7 +441,7 @@ void gPainter::setPalette(gRGB *colors, int start, int len)
 	gPalette *p=new gPalette;
 
 	o.parm.setPalette = new gOpcode::para::psetPalette;
-	p->data=new gRGB[len];
+	p->data=new gRGB[static_cast<size_t>(len)];
 
 	memcpy(static_cast<void*>(p->data), colors, len*sizeof(gRGB));
 	p->start=start;
